t/data/parsing/svg-1.dat

#data
abc<svg>foo<p>bbb</p>ccc</svg>xx<g></g>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     "abc"
|     <svg svg>
|       "foo"
|       <svg p>
|         "bbb"
|       "ccc"
|     "xx"
|     <g>

#data
abc<svg>fo<altglyph/>o</svg>bbb
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     "abc"
|     <svg svg>
|       "fo"
|       <svg altGlyph>
|       "o"
|     "bbb"

#data
<svg viewbox=abc><g startoffset=31 definitionurl=1 /></svg>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <svg svg>
|       viewBox="abc"
|       <svg g>
|         definitionurl="1"
|         startOffset="31"

#data
<svg xmlns="abc" xlink:tITle="bb" XML:lang="cc">aa</svg>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <svg svg>
|       xlink title="bb"
|       xml lang="cc"
|       xmlns xmlns="abc"
|       "aa"

#data
<svg><g xmlns="abc" xlink:tITle="bb" XML:lang="cc">aa</g></svg>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <svg svg>
|       <svg g>
|         xlink title="bb"
|         xml lang="cc"
|         xmlns xmlns="abc"
|         "aa"

#data
<svg><p>aaaa</svg>
#errors
1;13;m;not closed;;p
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <svg svg>
|       <svg p>
|         "aaaa"

#data
<svg>hoge<br>foo</svg>
#errors
1;17;m;not closed;;br
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <svg svg>
|       "hoge"
|       <svg br>
|         "foo"

#data
<svg><table><tr>foo
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <svg svg>
|       <svg table>
|         <svg tr>
|           "foo"

#data
<svg>foo<p>bar<p>bc</svg>
#errors
1;20;m;not closed;;p
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <svg svg>
|       "foo"
|       <svg p>
|         "bar"
|         <svg p>
|           "bc"

#data
<svg>fo<optgroup>bar<option>aa<optgroup>aa</svg>
#errors
1;43;m;not closed;;optgroup
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <svg svg>
|       "fo"
|       <svg optgroup>
|         "bar"
|         <svg option>
|           "aa"
|           <svg optgroup>
|             "aa"

#data
<svg>foo<html>BAR<body>abz</html>aaa</svg>
#errors
1;9;m;start tag not allowed;;html
1;18;m;start tag not allowed;;body
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <svg svg>
|       "fooBARabzaaa"

#data
<svg><title>foo<p>bar</p>&amp;</title></svg>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <svg svg>
|       <svg title>
|         "foo"
|         <p>
|           "bar"
|         "&"

#data
<svg><title>foo<p>bar&amp;</title></svg>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <svg svg>
|       <svg title>
|         "foo"
|         <p>
|           "bar&"

#data
<svg><xmp>foo<p>bar</p>&amp;</xmp></svg>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <svg svg>
|       <svg xmp>
|         "foo"
|         <svg p>
|           "bar"
|         "&"

#data
<svg><plaintext>foo<p>bar</p>&amp;</plaintext></svg>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <svg svg>
|       <svg plaintext>
|         "foo"
|         <svg p>
|           "bar"
|         "&"

#data
<svg><foreignobject>a<p>b</foreignobject>c</svg>d
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <svg svg>
|       <svg foreignObject>
|         "a"
|         <p>
|           "b"
|       "c"
|     "d"

#data
<svg><title>foo<svg><foo/></svg>hoge<p>bar</title></svg>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <svg svg>
|       <svg title>
|         "foo"
|         <svg svg>
|           <svg foo>
|         "hoge"
|         <p>
|           "bar"

#data
<svg><desc>foo<svg><foo/></svg>hoge<p>bar</desc></svg>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <svg svg>
|       <svg desc>
|         "foo"
|         <svg svg>
|           <svg foo>
|         "hoge"
|         <p>
|           "bar"

#data
<svg><title>foo<math><foo/></math>hoge<p>bar</title></svg>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <svg svg>
|       <svg title>
|         "foo"
|         <math math>
|           <math foo>
|         "hoge"
|         <p>
|           "bar"

#data
<svg><foreignobject><math><annotation-xml><svg><g/></svg></annotation-xml></math></foreignobject></svg>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <svg svg>
|       <svg foreignObject>
|         <math math>
|           <math annotation-xml>
|             <svg svg>
|               <svg g>

#data
<svg><FEdropshadow></fedropSHAdow></svg>a
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <svg svg>
|       <svg feDropShadow>
|     "a"
